Rm ckb-js-script
