--- examples/ckb-js-script/.gitignore
@@ -1,3 +1,0 @@
-/build
-/target
-/tests/failed_txs
