--- examples/ckb-js-script/js/hello.js
@@ -1,1 +1,0 @@
-console.log("hello, ckb-js-script!");
