+++ examples/ckb-ts-script/contracts/fib/tsconfig.base.json
@@ -1,0 +1,24 @@
+{
+  "compilerOptions": {
+    /* The QuickJS supports ES2022 */
+    "module": "ESNext",
+    /* Use esbuild to bundle the code */
+    "moduleResolution": "Bundler",
+    "target": "ES2022",
+    "incremental": true,
+    "allowJs": true,
+    "importHelpers": false,
+    "declaration": true,
+    "declarationMap": true,
+    "experimentalDecorators": true,
+    "useDefineForClassFields": false,
+    "esModuleInterop": true,
+    "strict": true,
+    "noImplicitAny": true,
+    "strictBindCallApply": true,
+    "strictNullChecks": true,
+    "alwaysStrict": true,
+    "noFallthroughCasesInSwitch": true,
+    "forceConsistentCasingInFileNames": true
+  }
+}
