+++ examples/ckb-ts-script/contracts/sudt/README.md
@@ -1,0 +1,10 @@
+TODO: Write this readme
+
+This project was bootstrapped with [`create-ckb-js-vm-app`](https://github.com/nervosnetwork/ckb-js-vm).
+
+## How to Run
+
+```bash
+pnpm build
+pnpm start
+```
